+++ mod-pubsub-server/src/main/java/org/folio/rest/impl/ModTenantAPI.java
@@ -4,35 +4,15 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
-import io.vertx.kafka.client.consumer.KafkaConsumer;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.folio.dao.util.LiquibaseUtil;
-import org.folio.kafka.KafkaConfig;
-import org.folio.kafka.PubSubConfig;
 import org.folio.rest.RestVerticle;
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.TenantAttributes;
-import org.folio.spring.SpringContextUtil;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.ws.rs.core.Response;
 import java.util.Map;
 
 public class ModTenantAPI extends TenantAPI {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(ModTenantAPI.class);
-  // stub event type needed to create a consumer for initial testing of kafka config,
-  // will be removed in scope of {@link https://issues.folio.org/browse/MODPUBSUB-46}
-  private static final String STUB_EVENT_TYPE = "record_created";
-
-  @Autowired
-  private KafkaConfig kafkaConfig;
-
-  public ModTenantAPI() {
-    SpringContextUtil.autowireDependencies(this, Vertx.currentContext());
-  }
 
   @Validate
   @Override
@@ -46,8 +26,6 @@ public class ModTenantAPI extends TenantAPI {
         vertx.executeBlocking(
           blockingFuture -> {
             LiquibaseUtil.initializeSchemaForTenant(vertx, tenantId);
-            // Create stub consumer
-            createKafkaConsumer(tenantId, vertx);
             blockingFuture.complete();
           },
           result -> handler.handle(postTenantAr)
@@ -55,27 +33,4 @@ public class ModTenantAPI extends TenantAPI {
       }
     }, context);
   }
-
-  private KafkaConsumer<String, String> createKafkaConsumer(String tenantId, Vertx vertx) {
-    PubSubConfig pubSubConfig = new PubSubConfig(tenantId, STUB_EVENT_TYPE);
-    Map<String, String> consumerProps = kafkaConfig.getConsumerProps();
-    consumerProps.put(ConsumerConfig.GROUP_ID_CONFIG, pubSubConfig.getGroupId());
-    return KafkaConsumer.<String, String>create(vertx, consumerProps)
-      .subscribe(pubSubConfig.getTopicName(), ar -> {
-        if (ar.succeeded()) {
-          LOGGER.info("Subscribed to topic {}", pubSubConfig.getTopicName());
-        } else {
-          LOGGER.error("Could not subscribe to topic", ar.cause());
-        }
-      })
-      .handler(record -> {
-        try {
-          String event = record.value();
-          LOGGER.info("Received event {}", event);
-        } catch (Exception e) {
-          LOGGER.error("Error reading event value", e);
-        }
-      });
-  }
-
 }