+++ mod-pubsub-server/src/main/java/org/folio/services/MessagingModuleService.java
@@ -1,11 +1,12 @@
 package org.folio.services;
 
 import io.vertx.core.Future;
-import org.folio.rest.util.MessagingModuleFilter;
 import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.MessagingModuleCollection;
 import org.folio.rest.jaxrs.model.PublisherDescriptor;
 import org.folio.rest.jaxrs.model.SubscriberDescriptor;
+import org.folio.rest.util.MessagingModuleFilter;
+import org.folio.rest.util.OkapiConnectionParams;
 
 /**
  * Messaging Module service
@@ -26,7 +27,7 @@ public interface MessagingModuleService {
    * before creating a new publisher
    *
    * @param publisherDescriptor publisher descriptor
-   * @param tenantId tenant id
+   * @param tenantId            tenant id
    * @return future with boolean
    */
   Future<Boolean> savePublisher(PublisherDescriptor publisherDescriptor, String tenantId);
@@ -45,10 +46,10 @@ public interface MessagingModuleService {
    * before creating a new subscriber
    *
    * @param subscriberDescriptor subscriber descriptor
-   * @param tenantId tenant id
+   * @param params               Okapi connection params
    * @return future with boolean
    */
-  Future<Boolean> saveSubscriber(SubscriberDescriptor subscriberDescriptor, String tenantId);
+  Future<Boolean> saveSubscriber(SubscriberDescriptor subscriberDescriptor, OkapiConnectionParams params);
 
   /**
    * Deletes module matching filter criteria