+++ mod-pubsub-server/src/main/java/org/folio/services/impl/KafkaPublisherServiceImpl.java
@@ -6,6 +6,7 @@ import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.kafka.client.producer.KafkaProducer;
 import io.vertx.kafka.client.producer.impl.KafkaProducerRecordImpl;
+import org.apache.commons.lang3.StringUtils;
 import org.folio.kafka.PubSubConfig;
 import org.folio.rest.jaxrs.model.AuditMessage;
 import org.folio.rest.jaxrs.model.AuditMessagePayload;
@@ -13,7 +14,7 @@ import org.folio.rest.jaxrs.model.Event;
 import org.folio.rest.util.MessagingModuleFilter;
 import org.folio.services.AuditMessageService;
 import org.folio.services.MessagingModuleService;
-import org.folio.services.PublishingService;
+import org.folio.services.PublisherService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -27,16 +28,21 @@ import static org.folio.rest.jaxrs.model.MessagingModule.ModuleRole.PUBLISHER;
 import static org.folio.rest.jaxrs.model.MessagingModule.ModuleRole.SUBSCRIBER;
 
 @Component
-public class KafkaPublishingServiceImpl implements PublishingService {
+public class KafkaPublisherServiceImpl implements PublisherService {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(KafkaPublishingServiceImpl.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(KafkaPublisherServiceImpl.class);
 
-  @Autowired
   private KafkaProducer<String, String> producer;
-  @Autowired
-  private AuditMessageService auditMessageService;
-  @Autowired
   private MessagingModuleService messagingModuleService;
+  private AuditMessageService auditMessageService;
+
+  public KafkaPublisherServiceImpl(@Autowired KafkaProducer<String, String> producer,
+                                   @Autowired MessagingModuleService messagingModuleService,
+                                   @Autowired AuditMessageService auditMessageService) {
+    this.producer = producer;
+    this.messagingModuleService = messagingModuleService;
+    this.auditMessageService = auditMessageService;
+  }
 
   @Override
   public Future<Boolean> publishEvent(Event event, String tenantId) {
@@ -67,7 +73,7 @@ public class KafkaPublishingServiceImpl implements PublishingService {
           LOGGER.error(errorMessage);
           saveAuditMessage(event, tenantId, AuditMessage.State.REJECTED);
           return Future.failedFuture(new BadRequestException(errorMessage));
-        } else if (!messagingModuleCollection.getMessagingModules().get(0).getActivated()) {
+        } else if (Boolean.FALSE.equals(messagingModuleCollection.getMessagingModules().get(0).getActivated())) {
           String error = format("Event type %s is not activated for tenant %s", event.getEventType(), tenantId);
           LOGGER.error(error);
           saveAuditMessage(event, tenantId, AuditMessage.State.REJECTED);
@@ -124,7 +130,7 @@ public class KafkaPublishingServiceImpl implements PublishingService {
   }
 
   private Future<Boolean> saveAuditMessagePayload(Event event, String tenantId) {
-    if (event.getEventPayload() != null) {
+    if (StringUtils.isNotEmpty(event.getEventPayload())) {
       return auditMessageService
         .saveAuditMessagePayload(new AuditMessagePayload()
           .withEventId(event.getId())