+++ mod-pubsub-server/src/test/java/org/folio/rest/impl/PublishTest.java
@@ -13,9 +13,13 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import java.time.LocalDate;
 import java.util.Collections;
 import java.util.UUID;
 
+import static java.lang.String.format;
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.hasItems;
 import static org.hamcrest.Matchers.is;
 
 @RunWith(VertxUnitRunner.class)
@@ -27,24 +31,38 @@ public class PublishTest extends AbstractRestTest {
     .withDescription("Created SRS Marc Bibliographic Record with order data in 9xx fields")
     .withEventTTL(1)
     .withSigned(false);
-  private static final String EVENT = new JsonObject()
+  private static final JsonObject EVENT = new JsonObject()
     .put("id", UUID.randomUUID().toString())
     .put("eventType", "record_created")
     .put("eventMetadata", new JsonObject()
       .put("tenantId", TENANT_ID)
       .put("eventTTL", 30)
-      .put("publishedBy", "mod-very-important-1.0.0"))
-    .put("eventPayload", "something very important").encode();
+      .put("publishedBy", "mod-very-important-1.0.0"));
 
   @Test
   public void shouldReturnBadRequestIfPublisherIsNotRegistered() {
     RestAssured.given()
       .spec(spec)
-      .body(EVENT)
+      .body(EVENT.encode())
       .when()
       .post(PUBLISH_PATH)
       .then()
       .statusCode(HttpStatus.SC_BAD_REQUEST);
+
+    LocalDate startDate = LocalDate.now().minusDays(1);
+    LocalDate endDate = LocalDate.now().plusDays(1);
+
+    String query = "?startDate=" + startDate.toString() + "&endDate=" + endDate.toString();
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(HISTORY_PATH + query)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("totalRecords", is(2))
+      .body("auditMessages.size()", is(2))
+      .body("auditMessages*.state", containsInAnyOrder("CREATED", "REJECTED"));
   }
 
   @Test
@@ -54,11 +72,26 @@ public class PublishTest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
-      .body(EVENT)
+      .body(EVENT.encode())
       .when()
       .post(PUBLISH_PATH)
       .then()
       .statusCode(HttpStatus.SC_BAD_REQUEST);
+
+    LocalDate startDate = LocalDate.now().minusDays(1);
+    LocalDate endDate = LocalDate.now().plusDays(1);
+
+    String query = "?startDate=" + startDate.toString() + "&endDate=" + endDate.toString();
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(HISTORY_PATH + query)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("totalRecords", is(2))
+      .body("auditMessages.size()", is(2))
+      .body("auditMessages*.state", containsInAnyOrder("CREATED", "REJECTED"));
   }
 
   @Test
@@ -69,11 +102,61 @@ public class PublishTest extends AbstractRestTest {
 
     RestAssured.given()
       .spec(spec)
-      .body(EVENT)
+      .body(EVENT.encode())
       .when()
       .post(PUBLISH_PATH)
       .then()
       .statusCode(HttpStatus.SC_NO_CONTENT);
+
+    LocalDate startDate = LocalDate.now().minusDays(1);
+    LocalDate endDate = LocalDate.now().plusDays(1);
+
+    String query = "?startDate=" + startDate.toString() + "&endDate=" + endDate.toString();
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(HISTORY_PATH + query)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("auditMessages*.state", hasItems("CREATED", "PUBLISHED"));
+  }
+
+  @Test
+  public void shouldPublishEventWithPayload() {
+    EventDescriptor eventDescriptor = postEventDescriptor(EVENT_DESCRIPTOR);
+    registerPublisher(eventDescriptor);
+    registerSubscriber(eventDescriptor);
+
+    RestAssured.given()
+      .spec(spec)
+      .body(EVENT.put("eventPayload", "something very important").encode())
+      .when()
+      .post(PUBLISH_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_NO_CONTENT);
+
+    LocalDate startDate = LocalDate.now().minusDays(1);
+    LocalDate endDate = LocalDate.now().plusDays(1);
+
+    String query = "?startDate=" + startDate.toString() + "&endDate=" + endDate.toString();
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(HISTORY_PATH + query)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("auditMessages*.state", hasItems("CREATED", "PUBLISHED"));
+
+    String eventId = EVENT.getString("id");
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(format(AUDIT_MESSAGES_PAYLOAD_PATH, eventId))
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("eventId", is(eventId));
   }
 
   private void registerPublisher(EventDescriptor eventDescriptor) {