+++ NEWS.md
@@ -21,6 +21,7 @@
 * Configured kafka client
 * Removed module table, modified messagingModule table, updated schemas
 * Created Publishing service
+* Created Consumer service
 
  | METHOD |             URL                                                                         | DESCRIPTION                                      |
  |--------|-----------------------------------------------------------------------------------------|--------------------------------------------------|